Documentation, removed non-standard features from query language.
Non-Standard features (also known as extensions to the standard) will be added in again in the future, in some form of extended query types.  However, for now, I want to focus on implementing the most common query features from standard SQL.  I'm not going to worry about breaking MySQL because of its lack of compliance with the SQL standard (or any other SQL, for that matter).