Add a limit and offset test
This is going to be a somewhat big deal, since I know that MySQL doesn't support the SQL standard when it comes to offsets (it uses 'LIMIT N' instead of 'FETCH (N) ROWS ONLY').  It looks like sqlite also doesn't support the SQL standard, based on my tests.  Time to fix that (and then clean up the tests).