Added some foreign key helpers
I've been running in to situations where I want to use an interface type as a field in a struct for foreign key purposes, but that makes database and input libraries which unmarshal to my struct type fail miserably, because they can't initialize the field.  I've added a couple of functions that should be the start of taking care of such things.