Add InSubQuery filter
In addition, we had to mess around with the bindVar logic, since the sub-query needs its variables bound based on the correct index of the containing query.  We're now using a bindVar placeholder string ("%s" for the time being), which is replaced by the actual bindVars when the query is run.